Close open socket file descriptors
To avoid leaking the open file descriptors and eventually crashing.

As long as the process is running, we have to make sure to close the
file descriptors we open. If we keep opening files without closing them,
eventually the process will run out of file descriptors, and opening the
next one will fail.

Our testing has revealed that xapi-nbd suffers from this issue - as
clients keep connecting, we do not close the corresponding sockets.
Eventually the NBD server crashes after a few hundred clients when it
runs out of file descriptors.

This code adds Lwt.finalize blocks after each socket is opened, to
ensure that they are closed, even if an exception gets raised. I have
also added Lwt.finalize blocks calling the Server.close and
channel.close functions, in addition to just closing the socket, because
these are from the nbd library and they may do more cleanup in the
future than just closing the socket, for example when TLS support will
be added. I have surrounded these calls with a try/catch block using the
ignore_exn helper, because an exception is raised when we double-close a
socket.

I've also extracted the with_session function to make the code more
readable, after adding all the finalizer blocks.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>