Time out after 5 minutes of client inactivity
Use a 5-minute timeout for input operations. This is effectively an
"inactivity timer": since the NBD server always keeps reading, as it
tries to read the next request after the current one has been processed,
this would achieve what we want - in case a client does not send a
request before the specified timeout, the read operation will fail.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>